Added Serial DMX Interface.  This is currently the skeleton code.
This has the ability to pipe DMX commands to any device at set
intervals depenant on the detected device.  Hopefully this will 
mean that the code can be re-used for multiple output devices.  
The next step will be to detect a FTDI chip (perhaps checking the 
VID is 0xEC70) and dump DMX output to it.


git-svn-id: https://qlc.svn.sourceforge.net/svnroot/qlc@733 0db28fee-5c31-0410-96ea-e69cce70e049