Initial implementation. This will currently let you add other functions to a shuffle and it will randomly select one at a fixed interval. There is also a blackoutState which could be selected.
This currently uses a new Function as it was easier to implement like that (it
is called every tick which afaict a VCWidget is not). I will refactor this to
use a VCWidget in the future.

I'm also aware that none of this is tested but while I'm not sure what the
correct behaviour is I find TDD hard. I will add tests when things stabilise.



git-svn-id: https://qlc.svn.sourceforge.net/svnroot/qlc@1226 0db28fee-5c31-0410-96ea-e69cce70e049